+++ src/main/nextflow/co2footprint/CO2FootprintPlugin.groovy
@@ -71,19 +71,21 @@ class CO2FootprintPlugin extends BasePlugin implements PluginAbstractExec {
         sessionTraceRecorder.stop()
         TraceRecord sessionRecord = sessionTraceRecorder.report()
 
-        CO2Record sessionCO2Record = observer.createCO2Record(sessionRecord)
-        observer.timeCiRecordCollector.stop()
+        if (observer != null) {
+            CO2Record sessionCO2Record = observer.createCO2Record(sessionRecord)
+            observer.timeCiRecordCollector.stop()
 
-        CO2RecordTree sessionStats = new CO2RecordTree(
-                observer.session?.runName,
-                [level: 'session'],
-                sessionCO2Record,
-                null,
-                [observer.workflowStats]
-        )
+            CO2RecordTree sessionStats = new CO2RecordTree(
+                    observer.session?.runName,
+                    [level: 'session'],
+                    sessionCO2Record,
+                    null,
+                    [observer.workflowStats]
+            )
 
-        // Render files
-        observer.renderFiles(sessionStats)
+            // Render files
+            observer.renderFiles(sessionStats)
+        }
 
         super.stop()
     }